+++ packages/fast-components-styles-msft/src/utilities/design-system.spec.ts
@@ -4,22 +4,25 @@ import defaultDesignSystem, {
 } from "../design-system";
 import * as designSystemUtils from "./design-system";
 
-Object.keys(designSystemUtils).forEach(
+
+Object.keys(designSystemUtils).filter((key: string) => key !== "getDesignSystemValue").forEach(
     (key: string): void => {
         describe(
             key,
             (): void => {
+                const designSystemKey: string = key === "getFontWeight" ? "fontWeight" : key;
+
                 test("should operate on design system defaults", (): void => {
                     expect(designSystemUtils[key]({} as DesignSystem)).toBe(
-                        defaultDesignSystem[key]
+                        defaultDesignSystem[designSystemKey]
                     );
                 });
                 test(`should return the ${key} property of the provided designSystem if it exists`, (): void => {
                     const customDesignSystem: DesignSystem = withDesignSystemDefaults({
-                        [key]: "custom value" as any,
+                        [designSystemKey]: "custom value" as any,
                     });
                     expect(designSystemUtils[key](customDesignSystem)).toBe(
-                        customDesignSystem[key]
+                        customDesignSystem[designSystemKey]
                     );
                 });
             }
